drop root_account_id from conversations tables
This removes the root_account_id from Conversation,
ConversationParticipant, ConversationMessage, and
ConversationMessageParticipant. We are opting to do something similar to
whats already being done in conversation with multiple root account ids
instead.

fixes VICE-549
flag=none

/ ---- ---- \
| Test Plan |
\ ---- ---- /

- Run the migration
  - bundle exec rake db:migrate:up VERSION=20200615170938
- Verify that the Conversation, ConversationParticipant,
  ConversationMessage, and ConversationMessageParticipant tables no
  longer have the `root_account_id` column

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240320
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>

+++ db/migrate/20200615170938_drop_root_account_id_from_conversations_tables.rb
@@ -14,18 +14,14 @@
 #
 # You should have received a copy of the GNU Affero General Public License along
 # with this program. If not, see <http://www.gnu.org/licenses/>.
-#
-class AddRootAccountIdToConversationMessageParticipants < ActiveRecord::Migration[5.2]
-  include MigrationHelpers::AddColumnAndFk
 
-  tag :predeploy
-  disable_ddl_transaction!
+class DropRootAccountIdFromConversationsTables < ActiveRecord::Migration[5.2]
+  tag :postdeploy
 
-  def up
-    add_column_and_fk :conversation_message_participants, :root_account_id, :accounts, if_not_exists: true
-  end
-
-  def down
-    remove_column :conversation_message_participants, :root_account_id
+  def change
+    remove_column :conversations, :root_account_id if column_exists?(:conversations, :root_account_id)
+    remove_column :conversation_participants, :root_account_id if column_exists?(:conversation_participants, :root_account_id)
+    remove_column :conversation_messages, :root_account_id if column_exists?(:conversation_messages, :root_account_id)
+    remove_column :conversation_message_participants, :root_account_id if column_exists?(:conversation_message_participants, :root_account_id)
   end
 end
